+++ backend/apps/dataetl/apps.py
@@ -1,6 +1,0 @@
-from django.apps import AppConfig
-
-
-class DataetlConfig(AppConfig):
-    default_auto_field = 'django.db.models.BigAutoField'
-    name = 'apps.dataetl'